move common function to resource file
