+++ lib/model/query/form-attachments.js
@@ -17,13 +17,10 @@ const { expectedFormAttachments } = require('../../data/schema');
 const { Frame, into } = require('../frame');
 const { Blob, Form } = require('../frames');
 const { unjoiner, insertMany } = require('../../util/db');
-const { ignoringResult, resolve, getOrNotFound, reject } = require('../../util/promise');
+const { ignoringResult, resolve } = require('../../util/promise');
 const { construct } = require('../../util/util');
 const Option = require('../../util/option');
-const { binary, contentDisposition, withEtag } = require('../../util/http');
-const { streamEntityCsvAttachment } = require('../../data/entity');
-const { md5sum } = require('../../util/crypto');
-const Problem = require('../../util/problem');
+
 
 ////////////////////////////////////////////////////////////////////////////////
 // CREATING ATTACHMENT SLOTS
@@ -144,32 +141,10 @@ select ${_unjoinMd5.fields} from form_attachments
   where "formDefId"=${formDefId}`)
   .then(map(_unjoinMd5));
 
-const streamAttachment = (attachment, response) => async ({ Blobs, Datasets, Entities }) => {
-  if (attachment.blobId == null && attachment.datasetId == null) {
-    return reject(Problem.user.notFound());
-  } else if (attachment.blobId != null) {
-    const blob = await Blobs.getById(attachment.blobId).then(getOrNotFound);
-    return withEtag(`"${blob.md5}"`, () => binary(blob.contentType, attachment.name, blob.content));
-  } else {
-    const dataset = await Datasets.getById(attachment.datasetId, true).then(getOrNotFound);
-    const properties = await Datasets.getProperties(attachment.datasetId);
-    const { lastEntity } = dataset.forApi();
-
-    const serverEtag = md5sum(lastEntity?.toISOString() ?? '1970-01-01');
-
-    return withEtag(serverEtag, async () => {
-      const entities = await Entities.streamForExport(attachment.datasetId);
-      response.append('Content-Disposition', contentDisposition(`${attachment.name}`));
-      response.append('Content-Type', 'text/csv');
-      return streamEntityCsvAttachment(entities, properties);
-    });
-  }
-};
 
 module.exports = {
   createNew, createVersion,
   update,
-  getAllByFormDefId, getByFormDefIdAndName,
-  getAllByFormDefIdForOpenRosa, streamAttachment
+  getAllByFormDefId, getByFormDefIdAndName, getAllByFormDefIdForOpenRosa
 };
