+++ lib/resources/forms.js
@@ -12,19 +12,45 @@ const { identity } = require('ramda');
 const { Blob, Form } = require('../model/frames');
 const { ensureDef } = require('../model/frame');
 const { QueryOptions } = require('../util/db');
-const { isTrue, xml, binary } = require('../util/http');
+const { isTrue, xml, binary, contentDisposition, withEtag } = require('../util/http');
 const Problem = require('../util/problem');
 const { sanitizeFieldsForOdata, setVersion } = require('../data/schema');
 const { getOrNotFound, reject, resolve, rejectIf } = require('../util/promise');
 const { success } = require('../util/http');
 const { formList, formManifest } = require('../formats/openrosa');
 const { noargs, isPresent, isBlank } = require('../util/util');
+const { streamEntityCsvAttachment } = require('../data/entity');
+const { md5sum } = require('../util/crypto');
 
 // excel-related util funcs/data used below:
 const isExcel = (contentType) => (contentType === 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet') || (contentType === 'application/vnd.ms-excel');
 const excelMimeTypes = {
   xls: 'application/vnd.ms-excel',
   xlsx: 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
+};
+
+const streamAttachment = async (container, attachment, response) => {
+  const { Blobs, Datasets, Entities } = container;
+
+  if (attachment.blobId == null && attachment.datasetId == null) {
+    return reject(Problem.user.notFound());
+  } else if (attachment.blobId != null) {
+    const blob = await Blobs.getById(attachment.blobId).then(getOrNotFound);
+    return withEtag(`${blob.md5}`, () => binary(blob.contentType, attachment.name, blob.content));
+  } else {
+    const dataset = await Datasets.getById(attachment.datasetId, true).then(getOrNotFound);
+    const properties = await Datasets.getProperties(attachment.datasetId);
+    const { lastEntity } = dataset.forApi();
+
+    const serverEtag = md5sum(lastEntity?.toISOString() ?? '1970-01-01');
+
+    return withEtag(serverEtag, async () => {
+      const entities = await Entities.streamForExport(attachment.datasetId);
+      response.append('Content-Disposition', contentDisposition(`${attachment.name}`));
+      response.append('Content-Type', 'text/csv');
+      return streamEntityCsvAttachment(entities, properties);
+    });
+  }
 };
 
 module.exports = (service, endpoint) => {
@@ -271,12 +297,14 @@ module.exports = (service, endpoint) => {
         .then((form) => auth.canOrReject('form.read', form))
         .then((form) => FormAttachments.getAllByFormDefId(form.def.id))));
 
-    service.get(`${base}/attachments/:name`, endpoint(({ FormAttachments, Forms }, { params, auth }, request, response) =>
-      getInstance(Forms, params)
+    service.get(`${base}/attachments/:name`, endpoint((container, { params, auth }, request, response) => {
+      const { FormAttachments, Forms } = container;
+      return getInstance(Forms, params)
         .then((form) => auth.canOrReject('form.read', form))
         .then((form) => FormAttachments.getByFormDefIdAndName(form.def.id, params.name)
           .then(getOrNotFound)
-          .then(attachment => FormAttachments.streamAttachment(attachment, response)))));
+          .then(attachment => streamAttachment(container, attachment, response)));
+    }));
   };
 
   // the linter literally won't let me break this apart..
@@ -406,13 +434,16 @@ module.exports = (service, endpoint) => {
       .then(checkFormToken(params.key))
       .then((form) => xml(form.xml))));
 
-  service.get('/test/:key/projects/:projectId/forms/:id/draft/attachments/:name', endpoint(({ FormAttachments, Forms }, { params }, request, response) =>
-    Forms.getByProjectAndXmlFormId(params.projectId, params.id, false, Form.DraftVersion)
+  service.get('/test/:key/projects/:projectId/forms/:id/draft/attachments/:name', endpoint((container, { params }, request, response) => {
+    const { FormAttachments, Forms } = container;
+    return Forms.getByProjectAndXmlFormId(params.projectId, params.id, false, Form.DraftVersion)
       .then(getOrNotFound)
       .then(ensureDef)
       .then(checkFormToken(params.key))
       .then((form) => FormAttachments.getByFormDefIdAndName(form.def.id, params.name)
         .then(getOrNotFound)
-        .then(attachment => FormAttachments.streamAttachment(attachment, response)))));
+        .then(attachment => streamAttachment(container, attachment, response)));
+  }));
+
 };
